+++ hockeypuck_rock/rockcraft.yaml
@@ -39,6 +39,7 @@ parts:
     organize:
       hockeypuck.conf: hockeypuck/etc/hockeypuck.conf
       delete_keys.py: hockeypuck/bin/delete_keys.py
+      rebuild_prefix_tree.py: hockeypuck/bin/rebuild_prefix_tree.py
       hockeypuck_wrapper.sh: hockeypuck/bin/hockeypuck_wrapper.sh
       migrate.sh: app/migrate.sh
   python:
